pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
license-files = ["LICENSE"]

[project]
name = "voxtream"
version = "0.2.0"
description = "Full-Stream Zero-shot TTS model with Extremely Low Latency and Speaking-rate Control"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Nikita Torgashov", email = "torgaschov.nikita@gmail.com" }
]
requires-python = ">=3.9"
keywords = [
  "text-to-speech",
  "streaming",
  "tts",
  "speech-synthesis",
  "voice-cloning"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Multimedia :: Sound/Audio",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
]

# install_requires (from requirements.txt)
dependencies = [
  "torch==2.4.0",
  "torchaudio==2.4.0",
  "torchtune==0.4.0",
  "torchao==0.9.0",
  "moshi==0.2.11",
  "transformers==4.50.0",
  "huggingface_hub==0.28.1",
  "g2p-en==2.1.0",
  "librosa==0.11.0",
  "soundfile==0.13.1",
  "inflect==7.5.0",
  "nltk==3.9.1",
  "gradio==4.44.1",
  "gradio_client==1.3.0",
  "pydantic==2.10.6",
  "setuptools>=70,<81",
  "silero-vad==6.2.0",
  "jiwer==4.0.0",
  "openai-whisper==20250625"
]

[project.optional-dependencies]
dev = ["black", "isort", "flake8", "mypy", "pytest"]

[project.urls]
Homepage = "https://herimor.github.io/voxtream2"
"Bug Reports" = "https://github.com/herimor/voxtream/issues"
Source = "https://github.com/herimor/voxtream"

[project.scripts]
voxtream = "voxtream.run:main"
voxtream-app = "voxtream.app:main"
voxtream-benchmark = "voxtream.benchmark:main"
voxtream-server = "voxtream.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["voxtream*"]

# Tool configs merged from your original pyproject.toml
[tool.black]
line-length = 88
target-version = ["py311"]
skip-string-normalization = false

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "W", "C90", "B", "I"]
lint.ignore = ["E501", "E402", "C901"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
implicit_optional = true
warn_unused_ignores = false
disable_error_code = [
  "arg-type",
  "assignment",
  "attr-defined",
  "call-overload",
  "import-untyped",
  "method-assign",
  "misc",
  "name-match",
  "no-redef",
  "operator",
  "override",
  "return-value",
  "union-attr",
  "var-annotated",
]
